Refactor: (remove duplications) centralize image picker functionalities with `ImagePickerWithLabel` component and update related forms
